transport-batch/requirements.txt
findspark==1.3.0     
pyspark==2.4.4
PyYAML==5.3
pymysql-0.9.3
